Changed the shake method to take a float instead of an int.  This was a bug that caused improper conversion.
Extended the startup wait time in the python file.

+++ shakr.pde
@@ -38,7 +38,7 @@ void setColor(float val)
 }
 
 //--- Set the shakr
-void shake(int val)
+void shake(float val)
 {
   // Turn on the lights and motors
   digitalWrite(vibPin, HIGH);
